refactor(platform-ui): extract iframe srcdoc builder to shared helper
Move the sandboxed-iframe document builder out of agent-output-review-panel
so the human-review task panel can render the same kind of inline HTML
report. The new helper also exports a typed resize-message guard, used by
the message listener in agent-output-review-panel and the upcoming task
context panel iframe.

No behaviour change.
